same for poets Shash Trevett is a poet, critic and a translator of Tamil poetry into English. Born in Sri Lanka, she now lives in the U.K. Her poetry has been widely published in journals and anthologies, she has read variously across the U.K and internationally and is a winner of a Northern Writers’ Award. Out of Sri Lanka: Tamil, Sinhala and English Poetry from Sri Lanka and its Diasporas (Bloodaxe 2023, Penguin India 2023) which she co-edited with Vidyan Ravinthiran and Seni Seneviratne, was a Times Literary Supplement’s Books of the Year for 2023. The Naming of Names, her first full collection, was published by Smith|Doorstop in September 2024 and her work has recently been conserved by the Poetry Archive. Shash has presented at academic conferences both in the U.K and abroad, and has been involved with the Singapore Translation Symposium, the PEN Translates awards and the Commonwealth Short Story Prize. She is a member of the Kinara Poetry Collective, is a Ledbury Critic and a Board Member of Modern Poetry in Translation.
Born and educated in Baddawi refugee camp in Lebanon, Yousif M. Qasmiyeh is a poet and scholar who completed his doctorate in English Literature at the University of Oxford. Time, the body and ruination inform his poetry and prose, which have appeared in Modern Poetry in Translation, Critical Quarterly, Cambridge Literary Review, PN Review, Stand, New England Review, Poetry London and Wasafiri. Yousif is the Creative Encounters Editor of the Migration and Society journal and Writer-in-Residence for the Refugee Hosts project. His debut collection, Writing the Camp (Broken Sleep Books, 2021), was a 2021 PBS Recommendation; was selected as one of the Best Poetry Books of 2021 by the Telegraph and the Irish Times; was highly commended by the 2021 Forward Prizes for Poetry; and was shortlisted for the 2022 Royal Society of Literature Ondaatje Prize. Yousif’s latest books are Eating the Archive (2023) and (with E. Fiddian-Qasmiyeh and Saiful Huq Omi) The Southern Eye (2024) (both published by Broken Sleep Books).